Bug fixes
The increment wasn't being applied to the variable because the method was being called without an operator to the variable.

The start displacement method was returning an incorrect value that wasn't an accurate displacement of start time. The time displacer also adds 1 hour indicating that the order starts being processed in the next hour. Deals with the minute value being discarded by simply 'rounding' to the nearest hour, even if it's 1 minute over the time.