Bug Fix
If Sunday is being skipped during the index reset, it wasn't being accounted in the dayDisplaced count. So each reset would only increment that days by a total of 6. Running a temporary variable to store the prior index value and compare to see if it's greater than the new index value - indicating a reset. Then it would increment the dayDisplaced count by 2 to account for the skipped Sunday index.